Refactor page tracking into the _app file for colocation
